app/requirements.txt
Flask>=0.12.2
python-slugify>=1.2.5
igdb-api-python>=1.0
python-dateutil>=2.7.2
Flask-Markdown>=0.3